Fix visual glitch immediately after triggering load
Modifying the state inside the KVO callback results in modifications to
the scrollview which in-turn results in more KVO callbacks. Once these
have all fired, the original KVO callback continues with old state. This
causes a visual glitch in the positioning of the view.

The solution is simply to move the state-change logic below the layout
logic.